+++ packages/commercetools/api-client/src/api/getShippingMethods/index.ts
@@ -1,4 +1,3 @@
-import { ApolloQueryResult } from 'apollo-client';
 import { apolloClient, getSettings } from '../../index';
 import defaultQuery from './defaultQuery';
 import { ShippingMethod } from './../../types/GraphQL';
@@ -7,9 +6,9 @@ interface ShippingMethodData {
   shippingMethods: ShippingMethod[];
 }
 
-const getShippingMethods = async (cartId?: string): Promise<ApolloQueryResult<ShippingMethodData>> => {
+const getShippingMethods = async (cartId?: string) => {
   const { acceptLanguage } = getSettings();
-  return await apolloClient.query({
+  return await apolloClient.query<ShippingMethodData>({
     query: defaultQuery,
     variables: { acceptLanguage, cartId },
     fetchPolicy: 'no-cache'