+++ packages/commercetools/api-client/src/index.ts
@@ -25,10 +25,6 @@ import createAccessToken from './helpers/createAccessToken';
 import { apiClientFactory } from '@vue-storefront/core';
 import { Config, ConfigurableConfig } from './types/setup';
 import { resolveCustomQueryVariables } from './helpers/search';
-export * from './types/Api';
-export * from './types/setup';
-export * as cartActions from './helpers/cart/actions';
-export * from './helpers/token';
 
 let apolloClient: ApolloClient<any> = null;
 
@@ -83,3 +79,9 @@ export {
   customerChangeMyPassword,
   customerUpdateMe
 };
+
+export * from './types/Api';
+export * from './types/GraphQL';
+export * from './types/setup';
+export * from './helpers/token';
+export * as cartActions from './helpers/cart/actions';