+++ packages/commercetools/composables/__tests__/useCart/useCart.spec.ts
@@ -24,6 +24,8 @@ jest.mock('@vue-storefront/core', () => ({
   })
 }));
 
+const customQuery = undefined;
+
 describe('[commercetools-composables] useCart', () => {
   beforeEach(() => {
     jest.clearAllMocks();
@@ -52,7 +54,7 @@ describe('[commercetools-composables] useCart', () => {
     const response = await addToCart({ currentCart: 'current cart', product: 'product1', quantity: 3 });
 
     expect(response).toEqual('some cart');
-    expect(apiAddToCart).toBeCalledWith('current cart', 'product1', 3, undefined);
+    expect(apiAddToCart).toBeCalledWith('current cart', 'product1', 3, customQuery);
   });
 
   it('creates a new cart and add an item', async () => {
@@ -62,7 +64,7 @@ describe('[commercetools-composables] useCart', () => {
     expect(loadCurrentCart).toBeCalled();
 
     expect(response).toEqual('some cart');
-    expect(apiAddToCart).toBeCalledWith('some cart', 'product1', 3, undefined);
+    expect(apiAddToCart).toBeCalledWith('some cart', 'product1', 3, customQuery);
   });
 
   it('removes from cart', async () => {
@@ -70,7 +72,7 @@ describe('[commercetools-composables] useCart', () => {
     const response = await removeFromCart({ currentCart: 'current cart', product: 'product1' });
 
     expect(response).toEqual('some cart');
-    expect(apiRemoveFromCart).toBeCalledWith('current cart', 'product1', undefined);
+    expect(apiRemoveFromCart).toBeCalledWith('current cart', 'product1', customQuery);
   });
 
   it('updates quantity', async () => {
@@ -82,7 +84,7 @@ describe('[commercetools-composables] useCart', () => {
     });
 
     expect(response).toEqual('some cart');
-    expect(apiUpdateCartQuantity).toBeCalledWith('current cart', { name: 'product1', quantity: 5 }, undefined);
+    expect(apiUpdateCartQuantity).toBeCalledWith('current cart', { name: 'product1', quantity: 5 }, customQuery);
   });
 
   it('clears cart', async () => {